add 400 error handling for deleteComment when an invalid comment_id is entered

+++ models/comments.js
@@ -1,5 +1,9 @@
 const db = require("../db/connection");
 
 exports.removeComment = (comment_id) => {
-  return db.query("DELETE FROM comments WHERE comment_id = $1;", [comment_id]);
+  if (!parseInt(comment_id)) {
+    return Promise.reject({ status: 400, msg: "Invalid comment ID" });
+  } else {
+    return db.query("DELETE FROM comments WHERE comment_id = $1;", [comment_id]);
+  }
 };
